requirements.txt
lxml
py-cord
python-dotenv
requests
beautifulsoup
